[ WIP ] Juani agregó un csv a spiffs para actualizar variables generales del dron ( configs ) sin tener que flashear de vuelta el programa. También agregó funciones para leer y procesar los datos del csv.

+++ Flight-Controller/components/bmi160/include/bmi160.h
@@ -82,7 +82,7 @@ typedef struct bmi160_t {
     i2c_params_t i2c;
 
     /** @brief Initialize Bmi160 object @param self: Address of Bmi160 object @param bmi_address: Bmi160 I2C address @param acc_mode: Accelerometer mode @param acc_freq: Accelerometer operation frequency @param acc_range: Accelerometer range @param gyro_mode: Gyroscope mode @param gyro_freq: Gyroscope operation frequency @param gyro_range: Gyroscope range @param gyro_offset_x: Gyroscope x offset @param gyro_offset_y: Gyroscope y offset @param gyro_offset_z: Gyroscope z offset @retval esp_err_t */
-    esp_err_t ( * init )( bmi160_t * self, int acc_mode,  int acc_freq,  int acc_range, int gyro_mode, int gyro_freq, int gyro_range, int gyro_offset_x, int gyro_offset_y, int gyro_offset_z );
+    esp_err_t ( * init )( bmi160_t * self, int acc_mode,  int acc_freq,  int acc_range, int gyro_mode, int gyro_freq, int gyro_range, float gyro_offset_x, float gyro_offset_y, float gyro_offset_z );
     esp_err_t ( * measure )( bmi160_t* self);
     esp_err_t ( * foc )( bmi160_t* self);
     
@@ -142,6 +142,6 @@ bmi160_t * Bmi160( int i2c_addr, int i2c_sda, int i2c_scl );
 esp_err_t bmi_init( bmi160_t * self,
     int acc_mode,  int acc_freq,  int acc_range,
     int gyro_mode, int gyro_freq, int gyro_range,
-    int gyro_offset_x, int gyro_offset_y, int gyro_offset_z
+    float gyro_offset_x, float gyro_offset_y, float gyro_offset_z
 );
